+++ apps/docs-chart/next.config.mjs
@@ -1,7 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  transpilePackages: ['@baneung-pack/ui', '@baneung-pack/chart'],
-};
-
-export default nextConfig;