+++ packages/chart/src/components/bar-chart-3d/index.ts
@@ -1,2 +1,0 @@
-export { BarChart3D } from './bar-chart-3d';
-export type { BarChart3DProps } from './bar-chart-3d';